Ransomware
It is a ransomware virus made with batch file Uses a tool Certutil to encryption It is in bat form Password of the compressed folder It is called, do not run it, otherwise the encryption will not be decrypted. You can create a file to decrypt it, and I will not put this file for you. Just reverse the code or put it in GPT chat to reflect it for you, and you can put the add step as entering a password. I prefer that you post my steps, such as making the password for the ID device of the computer or tablet to decrypt it